+++ Changelog.md
@@ -2,6 +2,8 @@
 
 - **Breaking Change** Renamed `verbose` in the NetworkLoggerPlugin to `isVerbose`.
 - **Breaking Change** `TargetType` now specifies its `ParameterEncoding`.
+- **Breaking Change** Removed unused `Moya.Error.data`.
+- `Moya.Error` now conforms to `LocalizedError` protocol.
 - Added documentation for `TargetType` and associated data structures.
 - Re-add `MultiTarget` to project.